# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * InstaVault API — Instagram media resolver and downloader
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
  ytDlpVersion?: string | null;
}

export interface ResolveRequest {
  /** Instagram URL (post, reel, story, highlight, profile, etc.) */
  url: string;
}

/**
 * Type of Instagram media
 */
export type MediaType = typeof MediaType[keyof typeof MediaType];


export const MediaType = {
  post: 'post',
  reel: 'reel',
  story: 'story',
  highlight: 'highlight',
  carousel: 'carousel',
  profile_pic: 'profile_pic',
  igtv: 'igtv',
} as const;

export interface MediaFormat {
  /** yt-dlp format selector string (passed verbatim to --format) */
  formatId: string;
  /** Human-readable label (e.g. "Best (1080p)", "720p", "Audio only") */
  label: string;
  /** File extension (mp4, jpg, mp3, etc.) */
  ext: string;
  /** Video codec (null or "none" for audio-only) */
  vcodec?: string | null;
  /** Audio codec */
  acodec?: string | null;
  /** Video height in pixels */
  height?: number | null;
  /** Video width in pixels */
  width?: number | null;
  /** Total bitrate in kbps */
  tbr?: number | null;
  /** Audio bitrate in kbps */
  abr?: number | null;
  /** Approximate file size in bytes */
  filesize?: number | null;
}

/**
 * Media type for this carousel item
 */
export type CarouselItemType = typeof CarouselItemType[keyof typeof CarouselItemType];


export const CarouselItemType = {
  photo: 'photo',
  video: 'video',
} as const;

export interface CarouselItem {
  /** Zero-based index in the carousel */
  index: number;
  /** Media type for this carousel item */
  type: CarouselItemType;
  /** Thumbnail URL for this item */
  thumbnailUrl?: string | null;
  /** Available download formats for this item */
  formats: MediaFormat[];
}

export interface MediaInfo {
  /** Instagram shortcode or media ID */
  id: string;
  type: MediaType;
  /** Instagram username of the author */
  author?: string | null;
  /** Display name of the author */
  authorDisplayName?: string | null;
  /** URL of the author's profile picture */
  authorAvatarUrl?: string | null;
  /** Thumbnail or preview image URL */
  thumbnailUrl?: string | null;
  /** Post caption text */
  caption?: string | null;
  /** Upload timestamp in ISO 8601 format */
  uploadedAt?: string | null;
  /** Video duration in seconds (null for photos) */
  durationSeconds?: number | null;
  /** Available formats for single-media posts (null for carousels) */
  formats?: MediaFormat[] | null;
  /** Individual items for carousel posts (null for single-media) */
  items?: CarouselItem[] | null;
}

export interface ErrorDetail {
  /** Machine-readable error code */
  code: string;
  /** Human-readable error message */
  message: string;
}

export interface ErrorResponse {
  error: ErrorDetail;
}

export type DownloadMediaParams = {
/**
 * Instagram URL of the media to download
 */
url: string;
/**
 * Format identifier from the resolve response (e.g. "best", "720p", "audio")
 */
formatId: string;
/**
 * For carousels, which item index to download (0-based)
 * @minimum 0
 */
index?: number;
};