# lib/api-zod/src/generated/types/carouselItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * InstaVault API — Instagram media resolver and downloader
 * OpenAPI spec version: 0.2.0
 */
import type { CarouselItemType } from './carouselItemType';
import type { MediaFormat } from './mediaFormat';

export interface CarouselItem {
  /** Zero-based index in the carousel */
  index: number;
  /** Media type for this carousel item */
  type: CarouselItemType;
  /** Thumbnail URL for this item */
  thumbnailUrl?: string | null;
  /** Available download formats for this item */
  formats: MediaFormat[];
}