# lib/api-zod/src/generated/types/errorDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * InstaVault API — Instagram media resolver and downloader
 * OpenAPI spec version: 0.2.0
 */

export interface ErrorDetail {
  /** Machine-readable error code */
  code: string;
  /** Human-readable error message */
  message: string;
}