# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * InstaVault API — Instagram media resolver and downloader
 * OpenAPI spec version: 0.2.0
 */
import type { ErrorDetail } from './errorDetail';

export interface ErrorResponse {
  error: ErrorDetail;
}