# lib/api-zod/src/generated/types/mediaFormat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * InstaVault API — Instagram media resolver and downloader
 * OpenAPI spec version: 0.2.0
 */

export interface MediaFormat {
  /** yt-dlp format selector string (passed verbatim to --format) */
  formatId: string;
  /** Human-readable label (e.g. "Best (1080p)", "720p", "Audio only") */
  label: string;
  /** File extension (mp4, jpg, mp3, etc.) */
  ext: string;
  /** Video codec (null or "none" for audio-only) */
  vcodec?: string | null;
  /** Audio codec */
  acodec?: string | null;
  /** Video height in pixels */
  height?: number | null;
  /** Video width in pixels */
  width?: number | null;
  /** Total bitrate in kbps */
  tbr?: number | null;
  /** Audio bitrate in kbps */
  abr?: number | null;
  /** Approximate file size in bytes */
  filesize?: number | null;
}