# lib/api-zod/src/generated/types/mediaInfo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * InstaVault API — Instagram media resolver and downloader
 * OpenAPI spec version: 0.2.0
 */
import type { CarouselItem } from './carouselItem';
import type { MediaFormat } from './mediaFormat';
import type { MediaType } from './mediaType';

export interface MediaInfo {
  /** Instagram shortcode or media ID */
  id: string;
  type: MediaType;
  /** Instagram username of the author */
  author?: string | null;
  /** Display name of the author */
  authorDisplayName?: string | null;
  /** URL of the author's profile picture */
  authorAvatarUrl?: string | null;
  /** Thumbnail or preview image URL */
  thumbnailUrl?: string | null;
  /** Post caption text */
  caption?: string | null;
  /** Upload timestamp in ISO 8601 format */
  uploadedAt?: string | null;
  /** Video duration in seconds (null for photos) */
  durationSeconds?: number | null;
  /** Available formats for single-media posts (null for carousels) */
  formats?: MediaFormat[] | null;
  /** Individual items for carousel posts (null for single-media) */
  items?: CarouselItem[] | null;
}