# lib/api-zod/src/generated/types/mediaType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * InstaVault API — Instagram media resolver and downloader
 * OpenAPI spec version: 0.2.0
 */

/**
 * Type of Instagram media
 */
export type MediaType = typeof MediaType[keyof typeof MediaType];


export const MediaType = {
  post: 'post',
  reel: 'reel',
  story: 'story',
  highlight: 'highlight',
  carousel: 'carousel',
  profile_pic: 'profile_pic',
  igtv: 'igtv',
} as const;