# lib/api-zod/src/generated/types/resolveRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * InstaVault API — Instagram media resolver and downloader
 * OpenAPI spec version: 0.2.0
 */

export interface ResolveRequest {
  /** Instagram URL (post, reel, story, highlight, profile, etc.) */
  url: string;
}